Add files via upload
I modified this version using vectors instead of arrays. The only large modifications I made were in the initialization definition of the vectors themselves, and function arguments while defining the function. The std::sort syntax is also different using vectors, I believe because vectors do not have to have a defined beginning and end so they do need to be specified.